build(fdp): remove built-in fdpclient to use new fairclient
