+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
   "click-option-group~=0.5.6",
   "sempyro >=1.4.0",
   "sparqlwrapper~=2.0",
+  "fairclient >= 1.0.0",
 ]
 
 [project.entry-points]
