+++ src/img2catalog/cli_app.py
@@ -4,6 +4,8 @@ from pathlib import Path
 import click
 import xnat
 from click_option_group import MutuallyExclusiveOptionGroup, optgroup
+from fairclient.fdpclient import FDPClient
+from fairclient.sparqlclient import FDPSPARQLClient
 from rdflib import URIRef
 
 from img2catalog import log
@@ -20,7 +22,6 @@ from img2catalog.const import (
     XNAT_USER_ENV,
     XNATPY_HOST_ENV,
 )
-from img2catalog.fdpclient import FDPClient, FDPSPARQLClient
 from img2catalog.xnat_parser import xnat_to_DCATDataset, xnat_to_FDP, xnat_to_RDF
 
 logger = logging.getLogger(__name__)
